+++ genet/libs/cudd/cudd/cuddUtil.c
@@ -2615,7 +2615,7 @@ Cudd_AverageDistance(
     double temeasured, nextmeasured;
     int i, j;
     int slots, nvars;
-    long long diff;
+    intptr_t diff;
     DdNode *scan;
     DdNodePtr *nodelist;
     DdNode *sentinel = &(dd->sentinel);
@@ -2638,13 +2638,13 @@ Cudd_AverageDistance(
 	for (j = 0; j < slots; j++) {
 	    scan = nodelist[j];
 	    while (scan != sentinel) {
-		diff = (long long) scan - (long long) cuddT(scan);
+		diff = (intptr_t) scan - (intptr_t) cuddT(scan);
 		tesubtotal += (double) ddAbs(diff);
-		diff = (long long) scan - (long long) Cudd_Regular(cuddE(scan));
+		diff = (intptr_t) scan - (intptr_t) Cudd_Regular(cuddE(scan));
 		tesubtotal += (double) ddAbs(diff);
 		temeasured += 2.0;
 		if (scan->next != sentinel) {
-		    diff = (long long) scan - (long long) scan->next;
+		    diff = (intptr_t) scan - (intptr_t) scan->next;
 		    nextsubtotal += (double) ddAbs(diff);
 		    nextmeasured += 1.0;
 		}
@@ -2663,7 +2663,7 @@ Cudd_AverageDistance(
 	scan = nodelist[j];
 	while (scan != NULL) {
 	    if (scan->next != NULL) {
-		diff = (long long) scan - (long long) scan->next;
+		diff = (intptr_t) scan - (intptr_t) scan->next;
 		nextsubtotal += (double) ddAbs(diff);
 		nextmeasured += 1.0;
 	    }
